Feat : 레스토랑 등록할 때 중복 검증 및 중복 저장 안되도록 구현

+++ src/domain/restaurantAPI.ts
@@ -5,8 +5,17 @@ const KEY = 'restaurant';
 const restaurantAPI = {
   save: (restaurant: RestaurantInfo) => {
     const json = window.localStorage.getItem(KEY);
-    const existingRestaurants = json ? JSON.parse(json) : [];
-    window.localStorage.setItem(KEY, JSON.stringify([...existingRestaurants, restaurant]));
+    const existingRestaurants: RestaurantInfo[] = json ? JSON.parse(json) : [];
+
+    const isExisting = existingRestaurants.some(
+      (existingRestaurant) => existingRestaurant.name === restaurant.name
+    );
+
+    if (!isExisting) {
+      window.localStorage.setItem(KEY, JSON.stringify([...existingRestaurants, restaurant]));
+    } else {
+      console.log(`${restaurant.name} 가 이미 존재합니다.`);
+    }
   },
 
   load: () => {
